attack_status_enum renamed to attack_utils
